+++ packages/chaingraph-nodes/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @badaitech/chaingraph-nodes
 
+## 0.2.0-dev.3
+
+### Patch Changes
+
+- feat: Add DirectSecret types for various API keys and update SecretNode type
+- Updated dependencies
+  - @badaitech/chaingraph-types@0.2.0-dev.3
+
 ## 0.2.0-dev.2
 
 ### Patch Changes